Allow `find(ObjectID)`. Fixes #99
