+++ test/options.test.js
@@ -1,16 +1,16 @@
 var should = require('should');
-var fo = require('../').find_options;
+var mongo = require('../');
+var fo = require('../lib/find_options.js');
 function find_options() {
   return fo(arguments);
 }
 
 describe('options tests', function () {
   var cb = function(){};
-  var noop = Boolean;
 
   it('should accept signature: empty arguments', function(){
     var options = find_options();
-    options.should.eql({ callback:noop, fields:{}, limit:0, query:{}, skip:0 });
+    options.should.eql({ callback:undefined, fields:{}, limit:0, query:{}, skip:0 });
   });
   it('should accept signature: "callback"', function(){
     var options = find_options(cb);
@@ -18,7 +18,12 @@ describe('options tests', function () {
   });
   it('should accept signature: "selector"', function(){
     var options = find_options({_id:"ABC123"});
-    options.should.eql({ callback:noop, fields:{}, limit:0, query:{_id:"ABC123"}, skip:0 });
+    options.should.eql({ callback:undefined, fields:{}, limit:0, query:{_id:"ABC123"}, skip:0 });
+  });
+  it('should handle ObjectIds', function(){
+    var id = mongo.ObjectID();
+    var options = find_options(id);
+    options.should.eql({ callback:undefined, fields:{}, limit:0, query:{_id:id}, skip:0 });
   });
   it('should accept signature: "selector, callback"', function(){
     var options = find_options({_id:"ABC123"}, cb);
@@ -26,7 +31,7 @@ describe('options tests', function () {
   });
   it('should accept signature: "selector, fields"', function(){
     var options = find_options({_id:"ABC123"}, {_id:-1});
-    options.should.eql({ callback:noop, fields:{_id:-1}, limit:0, query:{_id:"ABC123"}, skip:0 });
+    options.should.eql({ callback:undefined, fields:{_id:-1}, limit:0, query:{_id:"ABC123"}, skip:0 });
   });
   it('should accept signature: "selector, fields, callback"', function(){
     var options = find_options({_id:"ABC123"}, {_id:-1}, cb);
@@ -38,11 +43,11 @@ describe('options tests', function () {
   });
   it('should accept signature: "selector, options"', function(){
     var options = find_options({_id:"ABC123"}, {fields:{_id:-1}, skip:100});
-    options.should.eql({ callback:noop, fields:{_id:-1}, limit:0, query:{_id:"ABC123"}, skip:100 });
+    options.should.eql({ callback:undefined, fields:{_id:-1}, limit:0, query:{_id:"ABC123"}, skip:100 });
   });
   it('should accept signature: "selector, options"', function(){
     var options = find_options({_id:"ABC123"}, {projection:{_id:-1}, skip:100});
-    options.should.eql({ callback:noop, fields:{_id:-1}, limit:0, query:{_id:"ABC123"}, skip:100 });
+    options.should.eql({ callback:undefined, fields:{_id:-1}, limit:0, query:{_id:"ABC123"}, skip:100 });
   });
   it('should accept signature: "selector, options, callback"', function(){
     var options = find_options({_id:"ABC123"}, {fields:{_id:-1}, skip:100}, cb);
@@ -54,7 +59,7 @@ describe('options tests', function () {
   });
   it('should accept signature: "selector, fields, options"', function(){
     var options = find_options({_id:"ABC123"}, {_id:-1}, {skip:100});
-    options.should.eql({ callback:noop, fields:{_id:-1}, limit:0, query:{_id:"ABC123"}, skip:100 });
+    options.should.eql({ callback:undefined, fields:{_id:-1}, limit:0, query:{_id:"ABC123"}, skip:100 });
   });
   it('should accept signature: "selector, fields, options, callback"', function(){
     var options = find_options({_id:"ABC123"}, {_id:-1}, {skip:100}, cb);
@@ -62,7 +67,7 @@ describe('options tests', function () {
   });
   it('should accept signature: "selector, fields, skip, limit"', function(){
     var options = find_options({_id:"ABC123"}, {_id:-1}, 200, 100);
-    options.should.eql({ callback:noop, fields:{_id:-1}, limit:100, query:{_id:"ABC123"}, skip:200 });
+    options.should.eql({ callback:undefined, fields:{_id:-1}, limit:100, query:{_id:"ABC123"}, skip:200 });
   });
   it('should accept signature: "selector, fields, skip, limit, callback"', function(){
     var options = find_options({_id:"ABC123"}, {_id:-1}, 200, 100, cb);
@@ -70,7 +75,7 @@ describe('options tests', function () {
   });
   it('should accept signature: "selector, fields, skip, limit, timeout"', function(){
     var options = find_options({_id:"ABC123"}, {_id:-1}, 200, 100, 600000);
-    options.should.eql({ callback:noop, fields:{_id:-1}, limit:100, query:{_id:"ABC123"}, skip:200, timeout:600000 });
+    options.should.eql({ callback:undefined, fields:{_id:-1}, limit:100, query:{_id:"ABC123"}, skip:200, timeout:600000 });
   });
   it('should accept signature: "selector, fields, skip, limit, timeout, callback"', function(){
     var options = find_options({_id:"ABC123"}, {_id:-1}, 200, 100, 600000, cb);
